[ YC-977 ] Update lucene core to 8.x
- upgrade to 7.7.2
- totalHits is now a long
- replaced DocumentBoostFieldsScoreQuery with FunctionScoreQuery (requires reindexing after update)
- boost field are now indexed as DoubleDocValuesField

+++ search/search-lucene-core/src/main/java/org/yes/cart/search/dao/impl/GenericFTSLuceneImpl.java
@@ -185,7 +185,7 @@ public class GenericFTSLuceneImpl implements GenericFTS<Long, org.apache.lucene.
             }
             if (topDocs.totalHits > firstResult) {
 
-                lastResult = lastResult > topDocs.totalHits ? topDocs.totalHits : lastResult;
+                lastResult = lastResult > topDocs.totalHits ? (int) topDocs.totalHits : lastResult;
 
                 final List<Object[]> resItems = new ArrayList<>(lastResult - firstResult);
 
@@ -200,7 +200,7 @@ public class GenericFTSLuceneImpl implements GenericFTS<Long, org.apache.lucene.
                     logExplanation(searcher, query, sort, hit.doc);
                 }
 
-                return new Pair<>(resItems, topDocs.totalHits);
+                return new Pair<>(resItems, (int) topDocs.totalHits);
             }
         } catch (IllegalStateException ise) {
             LOG.warn("Failed to run query " + query + ", caused: " + ise.getMessage());
